+++ eden/Timer.h
@@ -1,0 +1,11 @@
+struct Timer {
+    timeval start;
+    Timer() {
+        gettimeofday(&start, 0);
+    }
+    double delta() {
+        timeval end;
+        gettimeofday(&end, 0);
+        return TimevalDeltaSec(start, end);
+    }
+};
